requirements-mps.txt
# Core
accelerate==1.9.0
einops==0.8.0
transformers==4.55.0
huggingface-hub==0.34.3
datasets==4.0.0
omegaconf==2.3.0
sentencepiece==0.2.0
opencv-python==4.12.0.88
matplotlib==3.10.5
onnxruntime>=1.22.1
peft==0.17.0
timm==1.0.19
qwen_vl_utils==0.0.11
loguru==0.7.3
diffusers @ git+https://github.com/huggingface/diffusers@7a2b78bf0f788d311cc96b61e660a8e13e3b1e63

# Needed by the script but missing from your list
pillow>=10.0.0
numpy>=1.23
tqdm>=4.66
safetensors>=0.4.3

# PyTorch for Apple Silicon (MPS). Install matching torch/vision pair.
# You can keep these unpinned and install with pip (PyTorch provides MPS in macOS wheels).
torch>=2.4; platform_system == "Darwin" and platform_machine == "arm64"
torchvision>=0.17; platform_system == "Darwin" and platform_machine == "arm64"

# Quantization lib (optional on MPS; harmless to keep installed)
optimum-quanto==0.2.7
